+++ Communication/Packets/Incoming/Rooms/Action/RemoveAllRightsEvent.cs
@@ -1,50 +1,16 @@
-﻿using Plus.Communication.Packets.Outgoing.Rooms.Engine;
-using Plus.Communication.Packets.Outgoing.Rooms.Permissions;
-using Plus.Communication.Packets.Outgoing.Rooms.Settings;
-using Plus.Database;
-using Plus.HabboHotel.GameClients;
+﻿using Plus.HabboHotel.GameClients;
 using Plus.HabboHotel.Rooms;
 
 namespace Plus.Communication.Packets.Incoming.Rooms.Action;
 
 internal class RemoveAllRightsEvent : RoomPacketEvent
 {
-    private readonly IRoomManager _roomManager;
-    private readonly IDatabase _database;
+    private readonly IRoomAccessService _roomAccessService;
 
-    public RemoveAllRightsEvent(IRoomManager roomManager, IDatabase database)
+    public RemoveAllRightsEvent(IRoomAccessService roomAccessService)
     {
-        _roomManager = roomManager;
-        _database = database;
+        _roomAccessService = roomAccessService;
     }
 
-    public override Task Parse(Room room, GameClient session, IIncomingPacket packet)
-    {
-        var instance = room;
-        if (!instance.CheckRights(session, true))
-            return Task.CompletedTask;
-        foreach (var userId in new List<int>(instance.UsersWithRights))
-        {
-            var user = instance.GetRoomUserManager().GetRoomUserByHabbo(userId);
-            if (user != null && !user.IsBot)
-            {
-                user.RemoveStatus("flatctrl 1");
-                user.UpdateNeeded = true;
-                user.GetClient()?.Send(new YouAreControllerComposer(0));
-            }
-            using (var dbClient = _database.GetQueryReactor())
-            {
-                dbClient.SetQuery("DELETE FROM `room_rights` WHERE `user_id` = @uid AND `room_id` = @rid LIMIT 1");
-                dbClient.AddParameter("uid", userId);
-                dbClient.AddParameter("rid", instance.Id);
-                dbClient.RunQuery();
-            }
-            session.Send(new FlatControllerRemovedComposer(instance, userId));
-            session.Send(new RoomRightsListComposer(instance));
-            session.Send(new UserUpdateComposer(instance.GetRoomUserManager().GetUserList().ToList()));
-        }
-        if (instance.UsersWithRights.Count > 0)
-            instance.UsersWithRights.Clear();
-        return Task.CompletedTask;
-    }
+    public override Task Parse(Room room, GameClient session, IIncomingPacket packet) => _roomAccessService.RemoveAllRights(room, session);
 }