+++ Communication/Packets/Incoming/Rooms/Action/RemoveRightsEvent.cs
@@ -1,49 +1,23 @@
-﻿using Plus.Communication.Packets.Outgoing.Rooms.Permissions;
-using Plus.Communication.Packets.Outgoing.Rooms.Settings;
-using Plus.Database;
-using Plus.HabboHotel.GameClients;
+﻿using Plus.HabboHotel.GameClients;
 using Plus.HabboHotel.Rooms;
 
 namespace Plus.Communication.Packets.Incoming.Rooms.Action;
 
 internal class RemoveRightsEvent : RoomPacketEvent
 {
-    private readonly IDatabase _database;
+    private readonly IRoomAccessService _roomAccessService;
 
-    public RemoveRightsEvent(IDatabase database)
+    public RemoveRightsEvent(IRoomAccessService roomAccessService)
     {
-        _database = database;
+        _roomAccessService = roomAccessService;
     }
 
     public override Task Parse(Room room, GameClient session, IIncomingPacket packet)
     {
-        if (!room.CheckRights(session, true))
-            return Task.CompletedTask;
         var amount = packet.ReadInt();
-        for (var i = 0; i < amount && i <= 100; i++)
-        {
-            var userId = packet.ReadInt();
-            if (userId > 0 && room.UsersWithRights.Contains(userId))
-            {
-                var user = room.GetRoomUserManager().GetRoomUserByHabbo(userId);
-                if (user != null && !user.IsBot)
-                {
-                    user.RemoveStatus("flatctrl 1");
-                    user.UpdateNeeded = true;
-                    user.GetClient()?.Send(new YouAreControllerComposer(0));
-                }
-                using (var dbClient = _database.GetQueryReactor())
-                {
-                    dbClient.SetQuery("DELETE FROM `room_rights` WHERE `user_id` = @uid AND `room_id` = @rid LIMIT 1");
-                    dbClient.AddParameter("uid", userId);
-                    dbClient.AddParameter("rid", room.Id);
-                    dbClient.RunQuery();
-                }
-                if (room.UsersWithRights.Contains(userId))
-                    room.UsersWithRights.Remove(userId);
-                session.Send(new FlatControllerRemovedComposer(room, userId));
-            }
-        }
-        return Task.CompletedTask;
+        var userIds = new List<int>(amount);
+        for (var i = 0; i < amount; i++)
+            userIds.Add(packet.ReadInt());
+        return _roomAccessService.RemoveRights(room, session, userIds);
     }
 }